# perf: `class_basename` helper