--- a/src/Illuminate/Support/helpers.php
+++ b/src/Illuminate/Support/helpers.php
@@ -70,7 +70,7 @@ if (! function_exists('class_basename')) {
     {
         $class = is_object($class) ? get_class($class) : $class;
 
-        return basename(str_replace('\\', '/', $class));
+        return (new \ReflectionClass($this))->getShortName();
     }
 }