The `class_basename` helper function can be updated for a ~50% performance gain.

Performance results on an Ubuntu server:
```
Reflection: 0.017999887466431s
Basename: 0.10418105125427s
Explode: 0.053033828735352s
Substring: 0.028043031692505s
```

The code I used to check performance (can be added to any `Eloquent` model):
```php
public function getClassReflection()
{
    return (new \ReflectionClass($this))->getShortName();
}

public function getClassBasename()
{
    return basename(str_replace('\\', '/', get_class($this)));
}

public function getClassExplode()
{
    $arr = explode('\\', get_class($this));
    return array_pop($arr);
}

public function getClassSubstring()
{
    $classNameWithNamespace = get_class($this);
    return substr($classNameWithNamespace, strrpos($classNameWithNamespace, '\\') + 1);
}

public static function benchmark()
{
    $model = new self();

    collect([
        "Reflection" => [],
        "Basename" => [],
        "Explode" => [],
        "Substring" => [],
    ])->map(function ($results, $strategy) use ($model) {
        for ($r = 0; $r < 10; $r++) { // 10 rounds [to get average]
            $start = microtime(true);
            $method = "getClass{$strategy}";
            for ($i = 0; $i < 100000; $i++) { // X100,000 [in each round]
                $model->{$method}();
            }
            $end = microtime(true);
            $results[] = ($end - $start);
            return $results;
        }
    })->each(function ($results, $strategy) {
        echo "$strategy: " . array_sum($results) / count($results) . "s\n";
    });
}
```