@GrahamCampbell 2021-10-02
Not only is this slower, but it's also a breaking change.