@GrahamCampbell 2021-10-02
The key thing to note here is that `basename` in PHP does not do any IO. It's just a dumb string function. The ReflectionClass stuff you are using will do IO, however...