@GrahamCampbell 2021-10-02
> but it's also a breaking change

The function currently works on classes that don't exist.