@GrahamCampbell 2021-10-02
To get accurate performance testing comparison, you should use a different class name each time. Then you will see that this is drastically slower, btw.